+++ src/module/apps/_module.mjs
@@ -1,9 +1,0 @@
-import CultureSheet from "./cultureSheet.mjs";
-import DestinySheet from "./destinySheet.mjs";
-import ManeuverSheet from "./maneuverSheet.mjs";
-
-export {
-  CultureSheet as Culture,
-  DestinySheet as Destiny,
-  ManeuverSheet as Maneuver
-};